+++ src/app/(home)/[org_name]/[repo_name]/page.tsx
@@ -1,9 +1,10 @@
+/* eslint-disable react-hooks/rules-of-hooks */
 "use client";
 import Pankuzu from "@/components/path/Pankuzu";
 import DataTableComponent from "@/components/DataTable/DataTableComponent";
 import ButtonArea from "@/app/(home)/components/ButtonArea";
 import LastPath from "@/components/path/LastPath";
-import { RepoContext } from "@/api/RepoContext";
+import { RepoContext } from "@/api/useRepoContext";
 import { useContext } from "react";
 const page = () => {
   const data = useContext(RepoContext);
