+++ src/app/(home)/[org_name]/page.tsx
@@ -1,31 +1,14 @@
+/* eslint-disable react-hooks/rules-of-hooks */
 "use client";
 import Pankuzu from "@/components/path/Pankuzu";
 import TableComponent from "@/components/TableComponent";
-import { useEffect, useContext } from "react";
 import ButtonArea from "@/app/(home)/components/ButtonArea";
 import DataTableComponent from "@/components/DataTable/DataTableComponent";
 import LastPath from "@/components/path/LastPath";
-import { RepoContext } from "@/api/RepoContext";
-import { fetchRepo } from "@/api/repository";
-import { repositoryAppInterface } from "@/api/interface/repository";
+import useOrgById from "@/hook/useOrgById";
 
 const page = () => {
-  const { data, setData } = useContext(RepoContext);
-
-  useEffect(() => {
-    const fetchData = async (org_id: string) => {
-      const res = await fetchRepo(org_id);
-      if (res !== null && res !== undefined) {
-        const data: repositoryAppInterface[] = res.apps;
-        setData(data);
-
-        return;
-      }
-    };
-    fetchData("test");
-  }, [data]);
-
-  console.log(data);
+  const data = useOrgById();
   const tadata = [
     {
       name: "Invoice 001",
@@ -40,6 +23,7 @@ const page = () => {
       deployments: 30,
     },
   ];
+
   return (
     <div className='container'>
       <Pankuzu />
