--- src/mock/TableData.ts
@@ -1,30 +1,0 @@
-export const TableData = {
-  header: ["invoice", "status", "amount", "dueDate"],
-
-  data: [
-    {
-      invoice: "001",
-      status: "Paid",
-      amount: "$300",
-      dueDate: "20/03/2023",
-    },
-    {
-      invoice: "002",
-      status: "Pending",
-      amount: "$500",
-      dueDate: "20/03/2023",
-    },
-    {
-      invoice: "003",
-      status: "Paid",
-      amount: "$300",
-      dueDate: "20/03/2023",
-    },
-    {
-      invoice: "004",
-      status: "Pending",
-      amount: "$500",
-      dueDate: "20/03/2023",
-    },
-  ],
-};
