+++ tsconfig.json
@@ -28,5 +28,8 @@
     ".next/types/**/*.ts",
     "src/components/SidebarData.tx"
   ],
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules"],
+  "eslint": {
+    "ignoreDuringBuilds": false
+  }
 }
